requirements.txt
streamlit
pandas
pyarrow
google-auth
google-api-python-client
pyyaml